+++ docs-app/tests/test-helper.ts
@@ -2,13 +2,12 @@ import { setApplication } from '@ember/test-helpers';
 import Application from 'docs-app/app';
 import config from 'docs-app/config/environment';
 import { setRunOptions } from 'ember-a11y-testing/test-support';
-import { start } from 'ember-qunit';
+import { setupEmberOnerrorValidation, start } from 'ember-qunit';
+import { loadTests } from 'ember-qunit/test-loader';
 import * as QUnit from 'qunit';
 import { setup } from 'qunit-dom';
 
 setApplication(Application.create(config.APP));
-
-setup(QUnit.assert);
 
 setRunOptions({
   rules: {
@@ -18,4 +17,7 @@ setRunOptions({
   },
 });
 
+setup(QUnit.assert);
+setupEmberOnerrorValidation();
+loadTests();
 start();
