+++ packages/ember-container-query/unpublished-development-types/index.d.ts
@@ -3,7 +3,6 @@
 
 import '@glint/environment-ember-loose';
 import '@glint/environment-ember-template-imports';
-import 'ember-source/types';
 
 import type EmberElementHelperRegistry from 'ember-element-helper/template-registry';
