+++ test-app/tsconfig.json
@@ -7,7 +7,10 @@
       "test-app/*": ["app/*"],
       "*": ["types/*"]
     },
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "types": [
+      "ember-source/types"
+    ]
   },
   "include": [
     "app/**/*",
